+++ server/app.js
@@ -25,7 +25,7 @@ app.get('/', (req, res) => {
 	res.sendFile(path.join(__dirname + '../index.html'))
 })
 
-const usersRoute = require('./routes/auth')
+const usersRoute = require('./routes/users')
 app.use('/auth', usersRoute)
 
 app.listen(PORT, () => console.log(`Listening on port ${PORT}`))
